Added final model
Spam reviews can directly impact the restaurant reputation, so accurately identifying them is important. The goal is high precision so when the model flags a review as spam, it is very likely to be spam (minimizing false positives). However, I don’t want recall to drop too much so the model still needs to catch a reasonable number of actual spam reviews. Occasionally detecting real reviews as spam is okay, but precision ensures that flagged reviews are truly spam most of the time.